+++ pkg/agent/authentication/authenticator/authenticator.go
@@ -1,21 +1,14 @@
-package auth
+package authenticator
 
 import (
 	"net/http"
-)
 
-type UserInfo struct {
-	Roles []string
-}
+	"github.com/spiffe/tornjak/pkg/agent/authentication/user"
+)
 
 type Authenticator interface {
 	// AuthenticateRequest takes request, verifies certain properties
 	// and returns relevant UserInfo to be interpreted by the Authorizer
 	// or error upon verification error
-	AuthenticateRequest(r *http.Request) (*UserInfo, error)
-}
-
-type Authorizer interface {
-	// Authorize Request
-	AuthorizeRequest(r *http.Request, u *UserInfo) error
+	AuthenticateRequest(r *http.Request) (*user.UserInfo, error)
 }
