+++ pkg/agent/authentication/user/user.go
@@ -1,0 +1,5 @@
+package user
+
+type UserInfo struct {
+	Roles []string
+}
